hostmem: add a host memory backend for Xen
Some virtual devices (e.g. NVDIMM) use the host memory backend to map
its backend resources to the guest. When those devices are used on Xen,
the mapping has to be managed out of QEMU. In order to reuse other parts
of the implementation of those devices, we introduce a host memory
backend for Xen (memory-backend-xen) which plays as a placeholder and
does not map any backend resource. Therefore, those devices when used on
Xen just needs to use the new backend.

Following is an example of the command options of an NVDIMM device using
a file backend on Xen:
   -object memory-backend-xen,id=mem1,mem-path=$FILENAME,size=$SIZE
   -device nvdimm,id=nvdimm1,memdev=mem1

Signed-off-by: Haozhong Zhang <haozhong.zhang@intel.com>