hostmem: add a host memory backend for Xen
Some virtual devices (e.g. NVDIMM) use the host memory backend to map
its backend resources to the guest. When those devices are used on Xen,
the mapping has to be managed out of QEMU. In order to reuse other parts
of the implementation of those devices, we introduce a host memory
backend for Xen (memory-backend-xen) which is mostly a placeholder
that only creates an empty memory region and does not actually
allocate and map the resource.

Following is an example of a vNVDIMM device backed by a host pmem
namespace,
   -object memory-backend-xen,id=mem1,mem-path=/dev/pmem0,size=4G
   -device nvdimm,id=nvdimm1,memdev=mem1