xen-hvm: create hotplug memory region on Xen
The current QEMU implementation allocates guest physical address for
vNVDIMM in the hotplug memory region, which was not created when QEMU
is used as Xen device model. Reuse existing code in i386/pc.c to
create the hotplug memory region for HVM domain on Xen, so that we can
reuse other parts of NVDIMM implementation with minor modifications.